Bijou64.jl

Bijou64.jl is a Julia implementation of the bijou64 variable-length integer encoding.

See the bijou64 specification for details.

For commentary on the motivations and performance outcomes, read the blog post announcing the encoding.

Installation

Bijou64.jl can be installed using Pkg.

pkg> add https://github.com/KyleSJohnston/Bijou64.jl

Example Usage

julia> using Bijou64

julia> Bijou64.encode(UInt64(67_000))
4-element Vector{UInt8}:
 0xfa
 0x00
 0x03
 0xc0

julia> Bijou64.decode(UInt64, [0xfa, 0x00, 0x03, 0xc0])
1-element Vector{UInt64}:
 0x00000000000105b8

API

Bijou64.Bijou64

Bijou64.BufferTooShort

Bijou64.decode

Bijou64.encode

Bijou64.encode

Bijou64.Bijou64 — Module

bijou64 variable-length integer encoding

Bijou64.decode — Method

decode(T, bytes)::Vector{T}

Decode bytes using the bijou64 variable-length integer encoding into a vector of unsigned integers of type T

T may be UInt8, UInt16, UInt32, or UInt64.

A BufferTooShort exception is thrown if bytes is empty or if bytes has too few elements for the tier of the encoded integer.

An OverflowError is thrown if there is an arithmetic overflow on tier 8 or if a tier of encoded integer is too large for T.

Bijou64.encode — Method

encode(v)

Encode v using the bijou64 variable-length integer encoding into a Vector{UInt8}

v must be a UInt8, a UInt16, a UInt32, or a UInt64.

Bijou64.encode — Method

encode(values)

Encode values using the bijou64 variable-length integer encoding into a Vector{UInt8}

values must be a Vector{T}, where T may be UInt8, UInt16, UInt32, or UInt64.

Bijou64.BufferTooShort — Type

Exception indicating that the byte buffer (vector) is too short